+++ infrastructure/images/dynamicsprepare.ps1
@@ -1,3 +1,0 @@
-Save-Dynamics365Resource -Resource Dynamics365Server90 -TargetDirectory C:\Install\Dynamics\Dynamics365Server90;
-Save-Dynamics365Resource -Resource Dynamics365Server90LanguagePackSve -TargetDirectory C:\Install\Dynamics\Dynamics365Server90LanguagePackSve;
-Install-Dynamics365Prerequisites;