+++ tests/PrepareMachine.ps1
@@ -1,6 +1,0 @@
-Save-Dynamics365Resource -Resource Dynamics365Server90 -TargetDirectory C:\Install\Dynamics\Dynamics365Server90
-Save-Dynamics365Resource -Resource Dynamics365Server90LanguagePackSve -TargetDirectory C:\Install\Dynamics\Dynamics365Server90LanguagePackSve
-#Save-Dynamics365Resource -Resource CRM2016 -TargetDirectory C:\Install\Dynamics\CRM2016
-#Save-Dynamics365Resource -Resource CRM2016LanguagePackSve -TargetDirectory C:\Install\Dynamics\CRM2016LanguagePackSve
-#Save-Dynamics365Resource -Resource CRM2016ServicePack2Update02 -TargetDirectory C:\Install\Dynamics\CRM2016ServicePack2Update02
-Install-Dynamics365Prerequisites